requirements/base.txt
Django==3.0.13
djangorestframework==3.12.0
safety==1.10.3
simple-settings==1.0.0
pytest==6.2.2
pytest-django==4.1.0
pytest-cov==2.11.1
ipython==7.21.0
psycopg2==2.8.6
dj_database_url==0.5.0
structlog==21.1.0